Cleaning up fix with queue.
