+++ docker-compose.tests.yml
@@ -7,12 +7,7 @@ services:
     command: /app/scripts/test-netcore.sh
     environment:
         TARANTOOL_REPLICATION_SOURCE: "tarantool:3301"
+        TEAMCITY_PROJECT_NAME: 123
     links:
         - tarantool
         - redis
-
-  redis:
-    image: redis:3.0-alpine
-    command: redis-server
-    ports:
-      - 6379:6379
