+++ docker-compose.yml
@@ -12,3 +12,9 @@ services:
         TARANTOOL_USER_NAME: admin
         TARANTOOL_USER_PASSWORD: adminPassword
         TARANTOOL_SLAB_ALLOC_ARENA: 0.1
+
+  redis:
+    image: redis:3.0-alpine
+    command: redis-server
+    ports:
+      - 6379:6379
