+++ src/progaudi.tarantool/IResponseWriter.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace ProGaudi.Tarantool.Client
+{
+    internal interface IResponseWriter : IDisposable
+    {
+        void BeginWriting();
+
+        bool IsConnected { get; }
+
+        Task Write(ArraySegment<byte> header, ArraySegment<byte> body);
+    }
+}
