+++ tests/progaudi.tarantool.tests/TestBase.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Threading.Tasks;
 using ProGaudi.Tarantool.Client.Model;
