+++ tests/progaudi.tarantool.tests/progaudi.tarantool.tests.csproj
@@ -24,8 +24,8 @@
     <PackageReference Include="StackExchange.Redis" Version="1.2.4" />
     <PackageReference Include="xunit" Version="2.2.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
-    <PackageReference Include="JetBrains.Annotations" Version="10.3.0" />
-    <PackageReference Include="Shouldly" Version="2.8.2" />
+    <PackageReference Include="JetBrains.Annotations" Version="10.4.0" />
+    <PackageReference Include="Shouldly" Version="2.8.3" />
     <ProjectReference Include="..\..\src\progaudi.tarantool\progaudi.tarantool.csproj" />
   </ItemGroup>
